remove exception and icmp types from classes
