--- IcmpTypes.py
@@ -1,5 +1,0 @@
-class IcmpTypes():
-    ECHO_REPLY = 0
-    DESTINATION_UNREACHABLE = 3
-    ECHO_REQUEST = 8
-    TIME_EXCEEDED = 11
