+++ exception.py
@@ -1,0 +1,10 @@
+class DestinationUnreachableException(Exception):
+    pass
+
+
+class ExceededTimeToLiveException(Exception):
+    pass
+
+
+class TimeoutException(Exception):
+    pass
