--- exception/TimeoutException.py
@@ -1,2 +1,0 @@
-class TimeoutException(Exception):
-    pass
