--- exception/__init__.py
@@ -1,5 +1,0 @@
-__all__ = [
-    'DestinationUnreachableException',
-    'ExceededTimeToLiveException',
-    'TimeoutException'
-]
